--- src/variables.js
@@ -1,6 +1,0 @@
-// Контейнеры для вставки карточек
-const filmsMainBlock = document.querySelector(`.films-list .films-list__container`);
-const filmsTopBlock = document.querySelector(`.rating-top .films-list__container`);
-const filmsCommentsBlock = document.querySelector(`.comments-top .films-list__container`);
-
-export {filmsMainBlock, filmsTopBlock, filmsCommentsBlock};
